+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Release Notes
 
+## WIP
+
+- Fix nav field
+
 ## 2.11.2 (2025-03-11)
 
 - Update content dashboard widget to sort models by latest updated item