+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Release Notes
 
-## WIP
+## 2.11.3 (2025-05-01)
 
 - Fix assets image api to support "original" value for width and height
 - Fix nav field